requirements.txt
keras-gpu
tensorflow-gpu
pandas
numpy
prody
multicoretsne
biovec
matplotlib
seaborn
opencv
scikit-learn
rdkit
tables
h5py
biopython
pydot
graphviz
hdbscan